mr-robot/pyproject.toml
[project]
name = "mr-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "redis>=5.0.0",
]
